Revert "Extract Listbox labels"
This reverts commit 59b27d116422696fbfa711943d9a16dc8df13d4a.
