+++ app/components/form/fields/ImageSelectField.tsx
@@ -11,9 +11,9 @@ import type { Image } from '@oxide/api'
 
 import type { InstanceCreateInput } from '~/forms/instance-create'
 import type { ListboxItem } from '~/ui/lib/Listbox'
-import { ImageLabel, selectedImageLabel } from '~/ui/lib/ListboxLabels'
+import { Slash } from '~/ui/lib/Slash'
 import { nearest10 } from '~/util/math'
-import { GiB } from '~/util/units'
+import { bytesToGiB, GiB } from '~/util/units'
 
 import { ListboxField } from './ListboxField'
 
@@ -52,11 +52,40 @@ export function BootDiskImageSelectField({
 }
 
 export function toListboxItem(i: Image, includeProjectSiloIndicator = false): ListboxItem {
+  const { name, os, projectId, size, version } = i
+  const formattedSize = `${bytesToGiB(size, 1)} GiB`
+
+  // filter out any undefined metadata and create a comma-separated list
+  // for the selected listbox item (shown in selectedLabel)
+  const condensedImageMetadata = [os, version, formattedSize].filter((i) => !!i).join(', ')
+  const metadataForSelectedLabel = condensedImageMetadata.length
+    ? ` (${condensedImageMetadata})`
+    : ''
+
+  // for metadata showing in the dropdown's options, include the project / silo indicator if requested
+  const projectSiloIndicator = includeProjectSiloIndicator
+    ? `${projectId ? 'Project' : 'Silo'} image`
+    : null
+  // filter out undefined metadata here, as well, and create a `<Slash />`-separated list
+  // for the listbox item (shown for each item in the dropdown)
+  const metadataForLabel = [os, version, formattedSize, projectSiloIndicator]
+    .filter((i) => !!i)
+    .map((i, index) => (
+      <span key={`${i}`}>
+        {index > 0 ? <Slash /> : ''}
+        {i}
+      </span>
+    ))
   return {
     value: i.id,
-    selectedLabel: selectedImageLabel(i),
+    selectedLabel: `${name}${metadataForSelectedLabel}`,
     label: (
-      <ImageLabel image={i} includeProjectSiloIndicator={includeProjectSiloIndicator} />
+      <>
+        <div>{name}</div>
+        <div className="text-tertiary selected:text-accent-secondary">
+          {metadataForLabel}
+        </div>
+      </>
     ),
   }
 }
